Add support for multiple hash algorithms.
Now the algorithm to be used in the execution can be specified in the configuration file.
The parameter will be read and using function pointers, the appropriate implementation will be used.

_blake3_hash* or _sha256_hash* functions will be called when calculating the hash of a file or directory.
Since the pointer is decided when the node is created, it does not generate any overhead every time the
functions are called during the lifetime of a node.